+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Updated
+1. Renamed _master_ branch to _main_.
+
 
 ## [0.8.8](https://github.com/uhppoted/uhppoted-lib/releases/tag/v0.8.8) - 2024-03-26